bigyellowjacket/server/requirements.txt
asyncio>=3.4.3
colorama>=0.4.4
websockets>=10.1
psutil>=5.8.0
pyOpenSSL>=20.0.1
aiodns>=3.0.0
cryptography>=3.4.7
aiohttp>=3.8.1
python-dateutil>=2.8.2
typing-extensions>=4.0.1
pydantic>=1.9.0
rich>=10.12.0